Main/Medium+MediumFormat+GuestOSType+SystemPropertiesImpl+Console+Global: consistently use bytes as size units, forgotten const value in API, MaxVDISize method renamed to InfoVDSize, STDMETHOD macro usage fixes, whitespace cleanup
Frontends/VirtualBox+VBoxManage+VBoxShell: adapt to changed disk size units

Main/StorageControllerImpl: check the storage controller instance limit to avoid creating unusable VM configs, simplify unnecessarily complex code for querying the controller properties


svn:sync-xref-src-repo-rev: r65913

+++ src/VBox/Main/include/Global.h
@@ -1,4 +1,4 @@
-/* $Id: Global.h 32120 2010-08-31 09:58:36Z noreply@oracle.com $ */
+/* $Id: Global.h 32531 2010-09-15 17:04:48Z klaus.espenlaub@oracle.com $ */
 
 /** @file
  *
@@ -60,7 +60,7 @@ public:
         const uint32_t                 osHint;
         const uint32_t                 recommendedRAM;
         const uint32_t                 recommendedVRAM;
-        const uint32_t                 recommendedHDD;
+        const uint64_t                 recommendedHDD;
         const NetworkAdapterType_T     networkAdapterType;
         const uint32_t                 numSerialEnabled;
         const StorageControllerType_T  dvdStorageControllerType;
